# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "DS Hospital API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/patient/login": {
      "post": {
        "tags": ["Patient"],
        "description": "Login for patients",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Patient's credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "john_doe"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful login",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "example": "john_doe"
                    },
                    "token": {
                      "type": "string",
                      "example": "<token_created>"
                    }
                  }
                },
                "date": {
                  "type": "string",
                  "example": "date_of_login"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/patient/logout": {
      "get": {
        "tags": ["Patient"],
        "description": "Logout for patients",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful logout",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "Patient Logged Out": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                },
                "date": {
                  "type": "string",
                  "example": "<date_of_logout>"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/patient/register": {
      "post": {
        "tags": ["Patient"],
        "description": "Register a new patient",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Patient's registration details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "john_doe"
                },
                "password": {
                  "type": "string",
                  "example": "password123"
                },
                "email": {
                  "type": "string",
                  "example": "john.doe@example.com"
                },
                "first_name": {
                  "type": "string",
                  "example": "John"
                },
                "last_name": {
                  "type": "string",
                  "example": "Doe"
                },
                "amka": {
                  "type": "string",
                  "example": "123456789"
                },
                "date_of_birth": {
                  "type": "string",
                  "example": "1997-20-06"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful registration",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string",
                      "example": "john_doe"
                    },
                    "token": {
                      "type": "string",
                      "example": "<token_created>"
                    }
                  }
                },
                "date": {
                  "type": "string",
                  "example": "<date_of_registration>"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/patient/book_appointment": {
      "post": {
        "tags": ["Patient"],
        "description": "Book an appointment with a doctor",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Appointment details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "appointment_date": {
                  "type": "string",
                  "example": "2024-07-01"
                },
                "appointment_time": {
                  "type": "string",
                  "example": "10:00"
                },
                "reason": {
                  "type": "string",
                  "example": "General Checkup"
                },
                "specialization": {
                  "type": "string",
                  "example": "Cardiologist"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment booked successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "cid": {
                      "type": "string",
                      "example": "APT_4746881"
                    },
                    "doctor_username": {
                      "type": "string",
                      "example": "dr_smith"
                    },
                    "patient_username": {
                      "type": "string",
                      "example": "john_doe"
                    },
                    "patient_name": {
                      "type": "string",
                      "example": "John"
                    },
                    "patient_surname": {
                      "type": "string",
                      "example": "Doe"
                    },
                    "doctor_name": {
                      "type": "string",
                      "example": "John"
                    },
                    "doctor_surname": {
                      "type": "string",
                      "example": "Smith"
                    },
                    "appointment_date": {
                      "type": "string",
                      "example": "2024-07-01"
                    },
                    "appointment_time": {
                      "type": "string",
                      "example": "10:00"
                    },
                    "reason": {
                      "type": "string",
                      "example": "General Checkup"
                    },
                    "cost": {
                      "type": "number",
                      "example": 200.0
                    },
                    "specialization": {
                      "type": "string",
                      "example": "Cardiologist"
                    }
                  }
                },
                "date": {
                  "type": "string",
                  "example": "date_of_book"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/patient/appointments": {
      "get": {
        "tags": ["Patient"],
        "description": "Fetch appointments for the logged-in patient",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of appointments",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "cid": {
                        "type": "string",
                        "example": "APT_4746881"
                      },
                      "doctor_name": {
                        "type": "string",
                        "example": "John"
                      },
                      "doctor_surname": {
                        "type": "string",
                        "example": "Smith"
                      },
                      "appointment_time": {
                        "type": "string",
                        "example": "10:00"
                      },
                      "appointment_date": {
                        "type": "string",
                        "example": "2024-07-01"
                      }
                    }
                  }
                },
                "date": {
                  "type": "string",
                  "example": "<date_of_request>"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/patient/appointment/{appointment_id}": {
      "get": {
        "tags": ["Patient"],
        "description": "Fetch details of a specific appointment",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "appointment_id",
            "in": "path",
            "description": "System generated cid for the appointment",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Details of the appointment",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "cid": {
                      "type": "string",
                      "example": "APT_4746881"
                    },
                    "doctor_name": {
                      "type": "string",
                      "example": "John"
                    },
                    "doctor_surname": {
                      "type": "string",
                      "example": "Smith"
                    },
                    "appointment_time": {
                      "type": "string",
                      "example": "10:00"
                    },
                    "appointment_date": {
                      "type": "string",
                      "example": "2024-07-01"
                    },
                    "specialization": {
                      "type": "string",
                      "example": "Cardiologist"
                    },
                    "cost": {
                      "type": "number",
                      "example": 200.0
                    },
                    "reason": {
                      "type": "string",
                      "example": "General Checkup"
                    }
                  }
                },
                "date": {
                  "type": "string",
                  "example": "<date_of_request>"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/patient/cancel_appointment/{appointment_id}": {
      "delete": {
        "tags": ["Patient"],
        "description": "Cancel a specific appointment",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "appointment_id",
            "in": "path",
            "description": "System generated cid for the appointment",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment canceled successfully",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "Appointment canceled": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                },
                "date": {
                  "type": "string",
                  "example": "<date_of_cancellation>"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
...
}